# docs/contracts/issuer.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SovereignID - Issuer API",
    "description": "Servicio emisor de Verifiable Credentials de la plataforma SovereignID. Vincula la wallet/DID emisor de una institucion y titulos emitidos a estudiantes usando la wallet primaria activa del estudiante. Los errores de negocio se devuelven como RFC 7807 Problem Details con un codigo estable en la extension `error`.",
    "contact": {
      "name": "SovereignID",
      "url": "https://github.com/nasotop/SovereignID"
    },
    "version": "v1"
  },
  "paths": {
    "/issuer/credentials/{credentialId}": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "Devuelve el detalle de una credencial autenticada si pertenece al titular del JWT.",
        "parameters": [
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderCredentialDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/issuer/credentials/{credentialId}/revoke": {
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Revokes an active credential after on-chain revocation.",
        "parameters": [
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeCredentialRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeCredentialRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialRevoked"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Comprueba que el servicio está operativo (liveness check).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/issuer/holders/me/credentials": {
      "get": {
        "tags": [
          "HolderCredentials"
        ],
        "summary": "Lista las credenciales del titular autenticado (JWT SIWE).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HolderCredentialSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/issuer/holders/me/credentials/{credentialId}": {
      "get": {
        "tags": [
          "HolderCredentials"
        ],
        "summary": "Devuelve el detalle de una credencial del titular autenticado.",
        "parameters": [
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderCredentialDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/issuer/institutions/{institutionId}/credentials": {
      "get": {
        "tags": [
          "InstitutionCredentials"
        ],
        "summary": "Lists credentials issued by an institution.",
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/issuer/institutions/{institutionId}/wallet": {
      "post": {
        "tags": [
          "InstitutionIssuerWallet"
        ],
        "summary": "Vincula la wallet/DID emisor de una institucion existente.",
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkInstitutionIssuerWalletRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkInstitutionIssuerWalletRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkInstitutionIssuerWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionIssuerWalletLinked"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/issuer/students/{studentId}/title": {
      "post": {
        "tags": [
          "StudentTitles"
        ],
        "summary": "Vincula un titulo emitido a un estudiante con wallet primaria activa.",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkStudentTitleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkStudentTitleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkStudentTitleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentTitleLinked"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CredentialRevoked": {
        "required": [
          "credentialId",
          "institutionId",
          "studentId",
          "status",
          "revokedAt",
          "revocationReason",
          "revocationTxHash"
        ],
        "type": "object",
        "properties": {
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "institutionId": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revocationReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "revocationTxHash": {
            "type": "string"
          }
        }
      },
      "CredentialSummary": {
        "required": [
          "credentialId",
          "institutionId",
          "studentId",
          "careerId",
          "credentialTypeCode",
          "subjectDid",
          "issuerDid",
          "status",
          "ipfsCid",
          "ipfsGatewayUrl",
          "contentHash",
          "transactionHash",
          "issuedAt",
          "revokedAt",
          "revocationReason",
          "studentLabel"
        ],
        "type": "object",
        "properties": {
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "institutionId": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "careerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "credentialTypeCode": {
            "type": "string"
          },
          "subjectDid": {
            "type": "string"
          },
          "issuerDid": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "ipfsCid": {
            "type": "string"
          },
          "ipfsGatewayUrl": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "transactionHash": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "revocationReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "studentLabel": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "HolderCredentialAnchors": {
        "required": [
          "ipfsCid",
          "ipfsGatewayUrl",
          "contentHash",
          "transactionHash",
          "blockNumber",
          "chainId",
          "eip712Signature"
        ],
        "type": "object",
        "properties": {
          "ipfsCid": {
            "type": "string"
          },
          "ipfsGatewayUrl": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "transactionHash": {
            "type": "string"
          },
          "blockNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chainId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "eip712Signature": {
            "type": "string"
          }
        }
      },
      "HolderCredentialDetail": {
        "required": [
          "id",
          "title",
          "issuerName",
          "typeCode",
          "status",
          "issuedAt",
          "expiresAt",
          "revokedAt",
          "subjectDid",
          "issuer",
          "anchors",
          "metadata"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "issuerName": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "revoked",
              "expired"
            ],
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "subjectDid": {
            "type": "string"
          },
          "issuer": {
            "$ref": "#/components/schemas/HolderCredentialIssuer"
          },
          "anchors": {
            "$ref": "#/components/schemas/HolderCredentialAnchors"
          },
          "metadata": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "HolderCredentialIssuer": {
        "required": [
          "did",
          "displayName",
          "code"
        ],
        "type": "object",
        "properties": {
          "did": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "HolderCredentialSummary": {
        "required": [
          "id",
          "title",
          "issuerName",
          "typeCode",
          "status",
          "issuedAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "issuerName": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "revoked",
              "expired"
            ],
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "InstitutionIssuerWalletLinked": {
        "required": [
          "institutionId",
          "walletAddress",
          "did",
          "publicKey"
        ],
        "type": "object",
        "properties": {
          "institutionId": {
            "type": "string",
            "format": "uuid"
          },
          "walletAddress": {
            "type": "string"
          },
          "did": {
            "type": "string"
          },
          "publicKey": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JsonElement": {},
      "LinkInstitutionIssuerWalletRequest": {
        "required": [
          "walletAddress",
          "did",
          "publicKey"
        ],
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string"
          },
          "did": {
            "type": "string"
          },
          "publicKey": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LinkStudentTitleRequest": {
        "required": [
          "credentialId",
          "careerId",
          "credentialTypeCode",
          "ipfsCid",
          "ipfsGatewayUrl",
          "contentHash",
          "transactionHash",
          "blockNumber",
          "chainId",
          "eip712Signature",
          "expiresAt",
          "metadata"
        ],
        "type": "object",
        "properties": {
          "credentialId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "careerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "credentialTypeCode": {
            "type": "string"
          },
          "ipfsCid": {
            "type": "string"
          },
          "ipfsGatewayUrl": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "transactionHash": {
            "type": "string"
          },
          "blockNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chainId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "eip712Signature": {
            "type": "string"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RevokeCredentialRequest": {
        "required": [
          "reason",
          "revocationTxHash",
          "blockNumber",
          "chainId",
          "eip712Signature",
          "revokedByUserId"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "revocationTxHash": {
            "type": "string"
          },
          "blockNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chainId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "eip712Signature": {
            "type": "string"
          },
          "revokedByUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "StudentTitleLinked": {
        "required": [
          "credentialId",
          "institutionId",
          "studentId",
          "careerId",
          "issuedToWalletId",
          "subjectDid",
          "issuerDid",
          "status",
          "issuedAt"
        ],
        "type": "object",
        "properties": {
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "institutionId": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "careerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "issuedToWalletId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectDid": {
            "type": "string"
          },
          "issuerDid": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT SIWE emitido por el servicio auth.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Credentials"
    },
    {
      "name": "Health"
    },
    {
      "name": "HolderCredentials"
    },
    {
      "name": "InstitutionCredentials"
    },
    {
      "name": "InstitutionIssuerWallet"
    },
    {
      "name": "StudentTitles"
    }
  ]
}